Move options to corresponding classes
